Add kaly scraper with tests
