+++ scraper-kalina/.coveragerc
@@ -1,0 +1,6 @@
+# .coveragerc to control coverage.py
+[run]
+omit =
+    */site-packages/*
+    */distutils/*
+    tests/*
