+++ scraper-kalina/.gitignore
@@ -1,0 +1,6 @@
+.coverage
+.pytest_cache
+coverage
+*posts.txt
+__pycache__
+scraper-env
