+++ scraper-kalina/main.py
@@ -1,12 +1,11 @@
-import scraper.py
+from module.url_scraper import UrlScraper
+from module.post_scraper import PostScraper
 
 if __name__ == '__main__':
     import time
     twenty_urls = UrlScraper.get_twenty_posts()
     post_scraper = PostScraper(twenty_urls)
-    posts = post_scraper.get_last_posts()
-    
     filename = str(time.time()) +  "posts.txt"
 
-    for p in posts:
+    for p in post_scraper.get_last_posts():
         p.print_to_file(filename)
